+++ .travis.yml
@@ -19,6 +19,10 @@ env:
   - JOB=other
   - JOB=javascript
 
+branches:
+  except:
+    - engine
+
 #matrix:
   #exclude:
     #- rvm: "1.9.3"
