+++ lib/katello/tasks/test.rake
@@ -1,31 +1,23 @@
 require File.expand_path("../engine", File.dirname(__FILE__))
 
 namespace :test do
-  namespace :katello do
-    task :test => ['db:test:prepare', 'test:katello:katello_test']
-    task :spec => ['db:test:prepare', 'test:katello:katello_spec']
 
-    Rake::TestTask.new(:katello_test) do |t|
-      t.libs << ["test", "#{Katello::Engine.root}/test"]
-      t.test_files = [
-        "#{Katello::Engine.root}/test/models/repository_test.rb"
-      ]
-      t.verbose = true
-    end
+  desc "Run the Katello plugin test suite."
+  task :katello => ['db:test:prepare'] do
+    # Set the test loader explicitly to ensure that the ci_reporter gem 
+    # doesn't override our test runner
+    ENV['TESTOPTS'] = "#{ENV['TESTOPTS']} #{Katello::Engine.root}/test/katello_test_runner.rb"
 
-    Rake::TestTask.new(:katello_spec) do |t|
+    test_task = Rake::TestTask.new('katello_test_task') do |t|
       t.libs << ["test", "#{Katello::Engine.root}/test", "spec", "#{Katello::Engine.root}/spec"]
       t.test_files = [
+        "#{Katello::Engine.root}/test/models/repository_test.rb",
         "#{Katello::Engine.root}/spec/models/activation_key_spec.rb"
       ]
       t.verbose = true
     end
-  end
 
-  desc "Test Katello plugin"
-  task 'katello' do
-    Rake::Task['test:katello:test'].invoke
-    Rake::Task['test:katello:spec'].invoke
+    Rake::Task[test_task.name].invoke
   end
 
 end
