+++ test/base_test_helper.rb
@@ -126,15 +126,6 @@ def configure_runcible
 end
 
 
-class ResourceTypeBackup
-  @@types_backup = ResourceType::TYPES.clone
-
-  def self.restore
-    ResourceType::TYPES.clear
-    ResourceType::TYPES.merge!(@@types_backup)
-  end
-end
-
 class CustomMiniTestRunner
   class Unit < MiniTest::Unit
