+++ test/katello_test_helper.rb
@@ -97,46 +97,3 @@ def disable_glue_layers(services=[], models=[], force_reload=false)
     FactoryGirl.reload
   end
 end
-
-class CustomMiniTestRunner
-  class Unit < MiniTest::Unit
-
-    def before_suites
-      # code to run before the first test
-      #configure_vcr
-    end
-
-    def after_suites
-      # code to run after the last test
-    end
-
-    def _run_suites(suites, type)
-      begin
-        if ENV['suite']
-          suites = suites.select do |suite|
-                     suite.name == ENV['suite']
-                   end
-        end
-
-        before_suites
-        super(suites, type)
-      ensure
-        after_suites
-      end
-    end
-
-    def _run_suite(suite, type)
-      begin
-        User.current = nil  #reset User.current
-        suite.before_suite if suite.respond_to?(:before_suite)
-        super(suite, type)
-      ensure
-        suite.after_suite if suite.respond_to?(:after_suite)
-        #ResourceTypeBackup.restore
-      end
-    end
-
-  end
-end
-
-MiniTest::Unit.runner = CustomMiniTestRunner::Unit.new
